Backend refactoring. Interface files moved
