+++ dpnp/backend/examples/example3.cpp
@@ -30,13 +30,13 @@
  * to calculate cos of input vector elements
  *
  * Possible compile line:
- * clang++ -g -fPIC examples/example3.cpp -Idpnp -Idpnp/backend -Ldpnp -Wl,-rpath='$ORIGIN'/dpnp -ldpnp_backend_c -o example3
+ * clang++ -g -fPIC examples/example3.cpp -Idpnp -Idpnp/backend/include -Ldpnp -Wl,-rpath='$ORIGIN'/dpnp -ldpnp_backend_c -o example3
  *
  */
 
 #include <iostream>
 
-#include "backend_iface.hpp"
+#include "dpnp_iface.hpp"
 
 int main(int, char**)
 {
