+++ dpnp/backend/include/dpnp_gen_1arg_1type_tbl.hpp
