+++ dpnp/backend/kernels/dpnp_krnl_fft.cpp
@@ -25,9 +25,9 @@
 
 #include <iostream>
 
-#include <backend_iface.hpp>
-#include "backend_fptr.hpp"
-#include "backend_utils.hpp"
+#include <dpnp_iface.hpp>
+#include "dpnp_fptr.hpp"
+#include "dpnp_utils.hpp"
 #include "queue_sycl.hpp"
 
 namespace mkl_dft = oneapi::mkl::dft;
