+++ dpnp/backend/kernels/dpnp_krnl_manipulation.cpp
@@ -27,10 +27,10 @@
 #include <iostream>
 #include <vector>
 
-#include <backend_iface.hpp>
+#include <dpnp_iface.hpp>
 
-#include "backend_fptr.hpp"
-#include "backend_utils.hpp"
+#include "dpnp_fptr.hpp"
+#include "dpnp_utils.hpp"
 #include "queue_sycl.hpp"
 
 template <typename _KernelNameSpecialization>
