+++ dpnp/backend/kernels/dpnp_krnl_sorting.cpp
@@ -25,8 +25,8 @@
 
 #include <iostream>
 
-#include <backend_iface.hpp>
-#include "backend_fptr.hpp"
+#include <dpnp_iface.hpp>
+#include "dpnp_fptr.hpp"
 #include "queue_sycl.hpp"
 
 template <typename _DataType, typename _idx_DataType>
