+++ dpnp/backend/src/queue_sycl.hpp
@@ -44,7 +44,7 @@
 #include <dpctl_sycl_queue_manager.h>
 #endif
 
-#include "backend_pstl.hpp" // this header must be included after <mkl.hpp>
+#include "dpnp_pstl.hpp" // this header must be included after <mkl.hpp>
 
 namespace mkl_rng = oneapi::mkl::rng;
